agent_rm query fix

+++ pkg/commands/agents_rm.go
@@ -5,10 +5,18 @@ import "github.com/pygrum/monarch/pkg/db"
 func cmdRm(names []string) {
 	var agents []db.Agent
 	if err := db.FindConditional("agent_id IN ?", names, &agents); err != nil {
-		if err = db.FindConditional("name IN ?", names, &agents); err != nil {
-			cLogger.Error("failed to find agent(s): %v", err)
+		cLogger.Error("failed to retrieve the specified agents: %v", err)
+		return
+	}
+	if len(agents) == 0 {
+		if err := db.FindConditional("name IN ?", names, &agents); err != nil {
+			cLogger.Error("failed to retrieve the specified agents: %v", err)
 			return
 		}
+	}
+	if len(agents) == 0 {
+		cLogger.Error("no agents with the provided names exist")
+		return
 	}
 	for _, agent := range agents {
 		if err := db.DeleteOne(&agent); err != nil {
